init react native with navigation
