+++ apps/rnative/src/screens/DefaultReactNativeScreen/components/Section.tsx
@@ -1,0 +1,35 @@
+import * as React from 'react';
+import {Text, useColorScheme, View} from 'react-native';
+import {Colors} from 'react-native/Libraries/NewAppScreen';
+import {styles} from './styles';
+
+type SectionProps = {
+  title?: string;
+  children?: React.ReactNode;
+};
+
+export const Section: React.FC<SectionProps> = ({children, title}) => {
+  const isDarkMode = useColorScheme() === 'dark';
+  return (
+    <View style={styles.sectionContainer}>
+      <Text
+        style={[
+          styles.sectionTitle,
+          {
+            color: isDarkMode ? Colors.white : Colors.black,
+          },
+        ]}>
+        {title}
+      </Text>
+      <Text
+        style={[
+          styles.sectionDescription,
+          {
+            color: isDarkMode ? Colors.light : Colors.dark,
+          },
+        ]}>
+        {children}
+      </Text>
+    </View>
+  );
+};
