+++ apps/rnative/src/screens/DefaultReactNativeScreen/components/styles.ts
@@ -1,0 +1,20 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  sectionContainer: {
+    marginTop: 32,
+    paddingHorizontal: 24,
+  },
+  sectionTitle: {
+    fontSize: 24,
+    fontWeight: '600',
+  },
+  sectionDescription: {
+    marginTop: 8,
+    fontSize: 18,
+    fontWeight: '400',
+  },
+  highlight: {
+    fontWeight: '700',
+  },
+});
