+++ apps/rnative/src/types/navigation-types.ts
@@ -1,0 +1,6 @@
+export type RootStackParamList = {
+  Home: undefined;
+  Details: undefined;
+};
+
+export type ScreenList = keyof RootStackParamList;
