+++ packages/ui-native/src/Button.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { Button as RNButton } from "react-native";
+import type { ButtonProps } from "react-native";
 
-export const Button = () => {
-  return <RNButton title="Cool Native " />;
+export const Button = ({ title = "Cool Native", ...props }: ButtonProps) => {
+  return <RNButton title={title} {...props} />;
 };
